// types.ts
export type CaseRarity = 'Very Common' | 'Common' | 'Uncommon' | 'Rare' | 'Very Rare';

// Represents the user's rarity selection preference, including the option for any rarity.
export type RaritySelection = CaseRarity | 'Any';

// Defines the available medical specialties. This enum populates the specialty selection dropdown on the HomeScreen.
// 'General Medicine' is a special case that triggers a random specialty selection in the case generation logic.
export enum Specialty {
  'General Medicine' = 'General Medicine', // This will function as the "Random" or "Any" category
  Cardiology = 'Cardiology',
  Pulmonology = 'Pulmonology',
  Gastroenterology = 'Gastroenterology',
  Neurology = 'Neurology',
  Dermatology = 'Dermatology',
  Endocrinology = 'Endocrinology',
  Nephrology = 'Nephrology',
  Oncology = 'Oncology',
  Rheumatology = 'Rheumatology',
  Psychiatry = 'Psychiatry',
  'Infectious Disease' = 'Infectious Disease',
  Hematology = 'Hematology',
  Orthopedics = 'Orthopedics',
}